+++ app/src/main/java/pl/lambada/songsync/ui/screens/about/components/AboutScreenTopBar.kt
@@ -1,0 +1,43 @@
+package pl.lambada.songsync.ui.screens.about.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MediumTopAppBar
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarScrollBehavior
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import pl.lambada.songsync.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AboutScreenTopBar(navController: NavController, scrollBehavior: TopAppBarScrollBehavior) {
+    MediumTopAppBar(
+        navigationIcon = {
+            IconButton(
+                onClick = {
+                    navController.popBackStack(navController.graph.startDestinationId, false)
+                }
+            ) {
+                Icon(
+                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                    contentDescription = stringResource(R.string.back),
+                )
+            }
+        },
+        title = {
+            Text(
+                modifier = Modifier.padding(start = 6.dp),
+                text = stringResource(id = R.string.about)
+            )
+        },
+        scrollBehavior = scrollBehavior
+    )
+}
